HMAN-7 refactored logic out into processor class
